CLI - update-users also create user-service.js and move UpdateUsersFromTwitch() into it
